setup.sh is updated.

+++ setup.sh
@@ -2,6 +2,11 @@
 
 set -e
 
+if [ -e package.json ]; then
+  echo "package.json already exists. Empty project is required to run this setup script."
+  exit 1
+fi
+
 printf '{}\n' > package.json
 npm i -D vite tailwindcss @tailwindcss/vite vite-plugin-handlebars fast-glob
 npm i alpinejs lucide swiper gsap
